requirements.txt
streamlit
pandas
matplotlib 
plotly
openpyxl
html5lib
PyMySQL
sqlalchemy
openai
python-dotenv